fix(es5): Corrected the exports definition so that the module is easier to use by regular node.js pr
As described at https://github.com/Microsoft/TypeScript/issues/2719, when the `export default`

construct is used, the consumers of es5 modules (like node.js) need to import the library using `var

describe = require('describe-module').default`. This is unnecessarily verbose and feels awkward.

However, when a module is exported using the `export =` syntax, the importing is much simpler and

all is needed is just the regular `var describe = require('describe-module');`